refactor(main): define new general structure
