+++ backend/cmd/orchestrator.go
@@ -1,0 +1,22 @@
+package main
+
+import (
+	"fmt"
+	"os"
+	"strings"
+)
+
+// Handle version flag
+func handleVersionFlag() {
+
+	if *versionFlag {
+		versionFile := "VERSION.txt"
+		versionData, err := os.ReadFile(versionFile)
+		if err == nil {
+			fmt.Println("Hyperloop UPV Backend Version:", strings.TrimSpace(string(versionData)))
+		} else {
+			fmt.Println("Hyperloop UPV Backend Version: unknown")
+		}
+		os.Exit(0)
+	}
+}
